Fix a security risk related to toggle a user's usage status for a package and reduced data leakage in the URL.
Previously the usage view simply accepted any user and package to toggle the given user's usage of the given package, but this allowed a malicious user to modify another user's package usage by visiting a carefully crafted URL. The usage view now no longer accepts a user argument; instead it always toggles the usage for the currently logged-in user.  In addition, usage now accepts a package slug instead of a package id, in order to reduce internal data leakage.