Removed the build_grids function in favor of simply using get_or_create in the edit element function.
This will save a lot of queries when someone adds a new package or feature to a grid (especially if there are a lot of elements that already exist), since we don't have to lookup (or create) all of elements at that time.  When the user goes to edit an element though, then we use get_or_create to make sure that they have an element to edit. Also added a sanity checks to ensure that we're only creating and retrieving elements for features and grid_packages that are a part of the same grid.