use Django template engine for more flexibility when rendering attributes
the prior method did a simple getattr on the package, but this doesn't allow
for any more powerful resolution of nested attributes. This would not allow
us to use more powerful linked data.

So for the pypi refactor, you could access the latest version with a template
variable like:

package.pypi.latest.version  >> 2.0.0

however passing an attribute like 'pypi.latest.version' fails massively with
the prior approach of a simple getattr. However Django has all the code needed
to resolve this into a proper value.

As a postscript, I'd rather see all this in python code, as some sort of utility
in the view code or something, but this change allows the pypi stuff to keep
moving forward.