:handshake: fix #1234: Add FavoritePackageRule to score favorite packages (#1237)
* fix #1234: Add FavoritePackageRule to score favorite packages

This new rule evaluates packages to determine if they are marked as favorite. It assigns a score based on the package's favorite status and provides relevant messages accordingly. This improvement helps refine the package scoring system.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix #1234: Add is_favorite property and update package favorite check

Introduced the is_favorite property to the package model to determine if a package is marked as a favorite based on favorite_count. Updated search rule to provide correct scoring based on this new property.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix #1234: Rename `is_favorite` to `has_favorite` in package model

The method name `is_favorite` was changed to `has_favorite` in the package model to better reflect the logic checking if a package has been marked as favorite by users. This change ensures consistency in method naming and improves code readability.

* fix #1234: Update score assignment in favorite package check

Previously, packages marked as favorites were assigned a zero score. This has been reversed so that favorite packages now receive the maximum score, while non-favorite packages are assigned a zero score. This aligns with the intended behavior of prioritizing favorite packages.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>